api/requirements.txt
# Core FastAPI and Server
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.20

# Image Processing
opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.1.3
pytesseract==0.3.13

# Barcode/QR Code Decoding
pyzbar==0.1.9

# Date Parsing
python-dateutil==2.9.0.post0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
playwright==1.48.0

# AI-Powered Search
tavily-python==0.5.0

# Caching (optional for now)
# redis==5.2.0

# Validation
pydantic==2.9.2
